WIP: fdre-a3xx/asm: handle integer consts

+++ fdre-a3xx/asm/ir-a3xx.c
@@ -587,7 +587,7 @@ struct ir3_attribute * ir3_attribute_create(struct ir3_shader *shader,
 }
 
 struct ir3_const * ir3_const_create(struct ir3_shader *shader,
-		int cstart, float v0, float v1, float v2, float v3)
+		int cstart, uint32_t v0, uint32_t v1, uint32_t v2, uint32_t v3)
 {
 	struct ir3_const *c = ir3_alloc(shader, sizeof(struct ir3_const));
 	c->val[0] = v0;
